Bring SurveyTally into load.py module so that we can use the same logger
