+++ .vscode/launch.json
@@ -87,18 +87,4 @@
       "args": ["-v", "1", "--spreadsheets", "--parent_dir", "2024", "--append_to_log_file"],
       //"args": ["-v", "1", "--compilation", "--last_n_days", "30"],
       "justMyCode": false
-    },
-    {
-      "name": "survey_tally",
-      "type": "debugpy",
-      "request": "launch",
-      "program": "${workspaceFolder}/smdb/scripts/survey_tally.py",
-      "console": "integratedTerminal",
-      //"args": ["-v", "1", "--parent_dir", "2024", "--read_xlsx"],
-      "args": ["-v", "2", "--parent_dir", "2024", "--write_csv"],
-      //"args": ["-v", "1", "--write_csv"],
-      //"args": ["-v", "1", "--parent_dir", "2022", "--read_xlsx", "--last_n_days", ".5"],
-      //"args": ["-v", "1", "--last_n_days", "0.5"],
     }
-  ]
-}
